[Installer] Create Xcode::TargetValidator
