+++ CHANGELOG.md
@@ -12,6 +12,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
   [Danielle Tomlinson](https://github.com/dantoml)
   [#5635](https://github.com/CocoaPods/CocoaPods/pull/5635)
 
+* Move Installer target verification into the Xcode namespace 
+  [Danielle Tomlinson](https://github.com/DanToml)
+  [#5607](https://github.com/CocoaPods/CocoaPods/pull/5607)
+
 ##### Bug Fixes
 
 * None.
